完成Chat模块的Record聊天的功能检测

+++ Config/src/main/java/com/seek/food/config/NacosConfig/Common/CommonParamRulesConfig.java
@@ -86,7 +86,7 @@ public class CommonParamRulesConfig {
         if (id==null||(id-idCapacity)>100*idCapacity||id<idCapacity) throw new BizException(ErrorCodeEnum.PARAM_ERROR);
     }
     public int getIdStart(long id){
-        return Math.toIntExact((int) id / idCapacity);
+        return Math.toIntExact(id / idCapacity);
     }
 
 }
